All remaining .c files renamed into .cc; fastcheck.py moved into ci folder
