+++ c/utils.cc
